made the DecisionPoint constructors protected to make them easier to override and fixed a minor compiler warning.
git-svn-id: https://europa-pso.googlecode.com/svn/PLASMA/trunk@2753 23e1bc29-d954-0410-9766-47de34158ba0

+++ CBPlanner/component/ConstrainedVariableDecisionPoint.cc
@@ -26,7 +26,7 @@ namespace EUROPA {
     check_error(m_var.isValid());
     if (m_var->lastDomain().isNumeric()) 
       if (!m_var->lastDomain().isFinite() || m_var->lastDomain().getSize() > 50) 
-	return m_choices[1] - m_choices[0];
+	return (unsigned int) (m_choices[1] - m_choices[0]);
     return m_choices.size();
   }
